Reformat log messages, switch from warn to err
These logs are followed by the return of an error value -- most likely,
a user of this library will follow this call by exiting the program. It
is better that an end user sees these messages marked as "error" rather
than "warning".

The messages themselves have been slightly reformatted to make them
easier to read for the end user.